+++ packages/db/src/config.ts
@@ -39,6 +39,7 @@ export const config: { [key: string]: Knex.Config } = {
     pool: {
       min: 2,
       max: 10,
+      acquireTimeoutMillis: 60000,
     },
   },
 
@@ -48,8 +49,9 @@ export const config: { [key: string]: Knex.Config } = {
       ? buildConnection(process.env.DATABASE_URL)
       : undefined,
     pool: {
-      min: 2,
+      min: 0,
       max: 20,
+      acquireTimeoutMillis: 60000,
     },
   },
 };